+++ efel/cppcore/LibV3.cpp
@@ -83,7 +83,7 @@ int LibV3::trace_check(mapStr2intVec& IntFeatureData,
   if (retval < 0) return -1;
 
   bool sane = true;
-  for (unsigned i = 0; i < peak_time.size(); i++) {
+  for (size_t i = 0; i < peak_time.size(); i++) {
     if (peak_time[i] < stim_start[0] || peak_time[i] > stim_end[0] * 1.05) {
       sane = false;
       break;
@@ -107,7 +107,7 @@ static int __peak_indices(double dThreshold, vector<double>& V,
   vector<int> upVec, dnVec;
   double dtmp;
   int itmp;
-  for (unsigned i = 1; i < V.size(); i++) {
+  for (size_t i = 1; i < V.size(); i++) {
     if (V[i] > dThreshold && V[i - 1] < dThreshold) {
       upVec.push_back(i);
     } else if (V[i] < dThreshold && V[i - 1] > dThreshold) {
@@ -120,7 +120,7 @@ static int __peak_indices(double dThreshold, vector<double>& V,
   }
   PeakIndex.clear();
   int j = 0;
-  for (unsigned i = 0; i < upVec.size(); i++) {
+  for (size_t i = 0; i < upVec.size(); i++) {
     dtmp = -1e9;
     itmp = -1;
     for (j = upVec[i]; j <= dnVec[i]; j++) {
@@ -185,14 +185,14 @@ int LibV3::ISI_values(mapStr2intVec& IntFeatureData,
 static int __ISI_CV(const vector<double>& isivalues, vector<double>& isicv) {
   // mean
   double isi_mean = 0.;
-  for (unsigned i = 0; i < isivalues.size(); i++) {
+  for (size_t i = 0; i < isivalues.size(); i++) {
     isi_mean += isivalues[i];
   }
   isi_mean /= isivalues.size();
 
   // sigma^2
   double variance = 0.;
-  for (unsigned i = 0; i < isivalues.size(); i++) {
+  for (size_t i = 0; i < isivalues.size(); i++) {
     double dev = isivalues[i] - isi_mean;
     variance += dev * dev;
   }
@@ -240,7 +240,7 @@ int LibV3::peak_voltage(mapStr2intVec& IntFeatureData,
   retVal = getDoubleVec(DoubleFeatureData, StringData, "V", V);
   if (retVal <= 0) return -1;
 
-  for (unsigned i = 0; i < PeakI.size(); i++) {
+  for (size_t i = 0; i < PeakI.size(); i++) {
     peakV.push_back(V[PeakI[i]]);
   }
   setDoubleVec(DoubleFeatureData, StringData, "peak_voltage", peakV);
@@ -267,7 +267,7 @@ int LibV3::firing_rate(mapStr2intVec& IntFeatureData,
   if (retVal <= 0) return -1;
 
   int nCount = 0;
-  for (unsigned i = 0; i < peakVTime.size(); i++) {
+  for (size_t i = 0; i < peakVTime.size(); i++) {
     if ((peakVTime[i] >= stimStart[0]) && (peakVTime[i] <= stimEnd[0])) {
       lastAPTime = peakVTime[i];
       nCount++;
@@ -299,7 +299,7 @@ int LibV3::peak_time(mapStr2intVec& IntFeatureData,
   retVal = getDoubleVec(DoubleFeatureData, StringData, "T", T);
   if (retVal <= 0) return -1;
 
-  for (unsigned i = 0; i < PeakI.size(); i++) {
+  for (size_t i = 0; i < PeakI.size(); i++) {
     pvTime.push_back(T[PeakI[i]]);
   }
   setDoubleVec(DoubleFeatureData, StringData, "peak_time", pvTime);
@@ -345,7 +345,7 @@ static int __spike_width1(const vector<double>& t, const vector<double>& v,
   vector<int> min_ahp_indices_plus(min_ahp_indices.size() + 1, start_index);
   copy(min_ahp_indices.begin(), min_ahp_indices.end(),
        min_ahp_indices_plus.begin() + 1);
-  for (unsigned i = 1; i < min_ahp_indices_plus.size(); i++) {
+  for (size_t i = 1; i < min_ahp_indices_plus.size(); i++) {
     double v_half = (v[peak_indices[i - 1]] + v[min_ahp_indices_plus[i]]) / 2.;
     // interpolate this one time step where the voltage is close to v_half in
     // the rising and in the falling edge
@@ -457,7 +457,7 @@ int LibV3::min_AHP_indices(mapStr2intVec& IntFeatureData,
   if (end_index > peak_indices_plus.back() + 5) {
     peak_indices_plus.push_back(end_index);
   }
-  for (unsigned i = 0; i < peak_indices_plus.size() - 1; i++) {
+  for (size_t i = 0; i < peak_indices_plus.size() - 1; i++) {
     int ahpindex = distance(
         v.begin(), min_element(v.begin() + peak_indices_plus[i],
                                v.begin() + peak_indices_plus[i + 1]));
@@ -552,7 +552,7 @@ static int __adaptation_index2(double StimStart, double StimEnd, double Offset,
   vector<double> ISI;
   // Select spike time between given time scale (stim_start and stim_end )
   // considet Offset also if it is given as input
-  for (unsigned i = 0; i < peakVTime.size(); i++) {
+  for (size_t i = 0; i < peakVTime.size(); i++) {
     if ((peakVTime[i] >= (StimStart - Offset)) &&
         (peakVTime[i] <= (StimEnd + Offset))) {
       SpikeTime.push_back(peakVTime[i]);
@@ -579,7 +579,7 @@ static int __adaptation_index2(double StimStart, double StimEnd, double Offset,
   // get addition and subtraction of ISIs
   double ISISum, ISISub, ADI;
   ADI = ISISum = ISISub = 0;
-  for (unsigned i = 1; i < ISI.size(); i++) {
+  for (size_t i = 1; i < ISI.size(); i++) {
     ISISum = ISI[i] + ISI[i - 1];
     ISISub = ISI[i] - ISI[i - 1];
     ADI = ADI + (ISISub / ISISum);
@@ -675,7 +675,7 @@ int LibV3::AP_amplitude(mapStr2intVec& IntFeatureData,
 
   vector<double> apamplitude;
   apamplitude.resize(peakvoltage.size());
-  for (unsigned i = 0; i < apamplitude.size(); i++) {
+  for (size_t i = 0; i < apamplitude.size(); i++) {
     apamplitude[i] = peakvoltage[i] - v[apbeginindices[i]];
   }
   setDoubleVec(DoubleFeatureData, StringData, "AP_amplitude", apamplitude);
@@ -704,7 +704,7 @@ static int __AP_width(const vector<double>& t, const vector<double>& v,
       find_if(t.begin(), t.end(), bind2nd(greater_equal<double>(), stimstart)));
   indices[0] = start_index;
   copy(minahpindices.begin(), minahpindices.end(), indices.begin() + 1);
-  for (unsigned i = 0; i < indices.size() - 1; i++) {
+  for (size_t i = 0; i < indices.size() - 1; i++) {
     /*
     // FWHM (not used):
     // half maximum
@@ -817,7 +817,7 @@ static int __AP_begin_indices(const vector<double>& t, const vector<double>& v,
       t.begin(),
       find_if(t.begin(), t.end(), bind2nd(greater_equal<double>(), stimstart)));
   minima.push_back(stimbeginindex);
-  for (unsigned i = 0; i < ahpi.size(); i++) {
+  for (size_t i = 0; i < ahpi.size(); i++) {
     if (ahpi[i] > stimbeginindex) {
       minima.push_back(ahpi[i]);
     }
@@ -833,7 +833,7 @@ static int __AP_begin_indices(const vector<double>& t, const vector<double>& v,
                                     bind2nd(greater_equal<double>(), stimend)));
     minima.push_back(stimendindex);
   }
-  for (unsigned i = 0; i < minima.size() - 1; i++) {
+  for (size_t i = 0; i < minima.size() - 1; i++) {
     // assure that the width of the slope is bigger than 4
     int newbegin = minima[i];
     int begin = minima[i];
@@ -908,7 +908,7 @@ static int __AP_end_indices(const vector<double>& t, const vector<double>& v,
   apei.resize(pi.size());
   vector<int> picopy(pi.begin(), pi.end());
   picopy.push_back(v.size() - 1);
-  for (unsigned i = 0; i < apei.size(); i++) {
+  for (size_t i = 0; i < apei.size(); i++) {
     // assure that the width of the slope is bigger than 4
     apei[i] = distance(
         dvdt.begin(),
@@ -950,7 +950,7 @@ int LibV3::AP_end_indices(mapStr2intVec& IntFeatureData,
 static int __AP_rise_indices(const vector<double>& v, const vector<int>& apbi,
                              const vector<int>& pi, vector<int>& apri) {
   apri.resize(apbi.size());
-  for (unsigned i = 0; i < apri.size(); i++) {
+  for (size_t i = 0; i < apri.size(); i++) {
     double halfheight = (v[pi[i]] + v[apbi[i]]) / 2.;
     vector<double> vpeak;
     vpeak.resize(pi[i] - apbi[i]);
@@ -997,7 +997,7 @@ static int __AP_fall_indices(const vector<double>& v, const vector<int>& apbi,
                              const vector<int>& apei, const vector<int>& pi,
                              vector<int>& apfi) {
   apfi.resize(apbi.size());
-  for (unsigned i = 0; i < apfi.size(); i++) {
+  for (size_t i = 0; i < apfi.size(); i++) {
     double halfheight = (v[pi[i]] + v[apbi[i]]) / 2.;
     vector<double> vpeak(&v[pi[i]], &v[apei[i]]);
     transform(vpeak.begin(), vpeak.end(), vpeak.begin(),
@@ -1046,7 +1046,7 @@ static int __AP_duration(const vector<double>& t,
                          const vector<int>& endindices,
                          vector<double>& apduration) {
   apduration.resize(apbeginindices.size());
-  for (unsigned i = 0; i < apduration.size(); i++) {
+  for (size_t i = 0; i < apduration.size(); i++) {
     apduration[i] = t[endindices[i]] - t[apbeginindices[i]];
   }
   return apduration.size();
